+++ source/unless.js
@@ -1,7 +1,11 @@
-export function unless(predicate, whenFalse){
-  if (arguments.length === 1){
-    return _whenFalse => unless(predicate, _whenFalse)
-  }
+import { curry } from './curry.js'
 
-  return input => predicate(input) ? input : whenFalse(input)
+function unlessFn(
+  predicate, whenFalseFn, input
+){
+  if (predicate(input)) return input
+
+  return whenFalseFn(input)
 }
+
+export const unless = curry(unlessFn)